"factored out SPI and UART from HAL; updates #1173"
