+++ software/controller/lib/hal/hal_stm32_regs.h
@@ -69,175 +69,6 @@ struct InterruptControlStruct {
 typedef volatile InterruptControlStruct InterruptControlReg;
 inline InterruptControlReg *const NvicBase = reinterpret_cast<InterruptControlReg *>(0xE000E100);
 
-// [RM] 38.8 USART Registers (pg 1238)
-struct UartStruct {
-  union {
-    struct {
-      uint32_t enable : 1;
-      uint32_t stop_mode : 1;
-      uint32_t rx_enable : 1;
-      uint32_t tx_enable : 1;
-      uint32_t idle_interrupt : 1;
-      uint32_t rx_interrupt : 1;
-      uint32_t tx_complete_interrupt : 1;
-      uint32_t tx_interrupt : 1;
-      uint32_t parity_error_interrupt : 1;
-      uint32_t parity : 1;
-      uint32_t parity_control_enable : 1;
-      uint32_t rx_wakeup_method : 1;
-      uint32_t word_length0 : 1;
-      uint32_t mute_mode_enable : 1;
-      uint32_t char_match_interrupt : 1;
-      uint32_t oversampling_mode : 1;
-      uint32_t driver_deassertion_time : 5;  // Driver Enable de-assertion time
-      uint32_t driver_assertion_time : 5;    // Driver Enable assertion time
-      uint32_t rx_timeout_interrupt : 1;
-      uint32_t eob_interrupt : 1;  // End of Block interrupt enable
-      uint32_t word_length1 : 1;
-      uint32_t reserved : 3;
-    } bitfield;
-    uint32_t full_reg;
-  } control_reg1;  // Control Register 1 (USART_CR1) [RM] 38.8.1 (pg 1238)
-  union {
-    struct {
-      uint32_t reserved2 : 4;
-      uint32_t address_detection_mode : 1;  // 7-bit | 4bit Address Detection
-      uint32_t break_detection_length : 1;
-      uint32_t break_detection_interrput : 1;
-      uint32_t reserved1 : 1;
-      uint32_t last_bit_clock_pulse : 1;
-      uint32_t clock_phase : 1;
-      uint32_t clock_polarity : 1;
-      uint32_t clock_enable : 1;
-      uint32_t stop_bits : 2;
-      uint32_t lin_mode_enable : 1;
-      uint32_t swap_rx_tx_pins : 1;
-      uint32_t rx_inversion : 1;
-      uint32_t tx_inversion : 1;
-      uint32_t data_inversion : 1;
-      uint32_t msb_first : 1;
-      uint32_t auto_baudrate_enable : 1;
-      uint32_t auto_baudrate_mode : 2;
-      uint32_t rx_timeout_enable : 1;
-      uint32_t addr : 8;  // used for character detection during normal
-                          // reception.  This bit field can only be written
-                          // when reception is disabled (RE = 0) or the
-                          // USART is disabled (UE=0)
-    } bitfield;
-    uint32_t full_reg;
-  } control2;  // Control Register 2 (USART_CR2) [RM] 38.8.2 (pg 1241)
-  union {
-    struct {
-      uint32_t error_interrupt : 1;
-      uint32_t irda_mode : 1;
-      uint32_t irda_low_power : 1;
-      uint32_t half_duplex : 1;
-      uint32_t smartcard_nack_enable : 1;
-      uint32_t smartcard_mode : 1;
-      uint32_t rx_dma : 1;
-      uint32_t tx_dma : 1;
-      uint32_t rts_enable : 1;
-      uint32_t cts_enable : 1;
-      uint32_t cts_interrupt : 1;
-      uint32_t onebit : 1;  // One sample bit method enable
-      uint32_t overrun_disable : 1;
-      uint32_t dma_disable_on_rx_error : 1;
-      uint32_t driver_enable_mode : 1;
-      uint32_t driver_enable_polarity : 1;
-      uint32_t reserved2 : 1;
-      uint32_t smartcard_auto_retry_count : 3;
-      uint32_t wakeup_stop_interrupt_select : 2;
-      uint32_t wakeup_stop_interrupt_enable : 1;
-      uint32_t uart_clock_stop_mode : 1;
-      uint32_t tx_complete_before_guard_interrupt : 1;
-      uint32_t reserved : 7;
-    } bitfield;
-    uint32_t full_reg;
-  } control3;  // Control Register 3 [RM] 38.8.3  (pg 1245)
-  uint32_t baudrate;
-  uint32_t guard_time;
-  union {
-    struct {
-      uint32_t rx_timeout : 24;
-      uint32_t block_length : 8;
-    } bitfield;
-    uint32_t full_reg;
-  } timeout;  // Receiver Timeout Register (USART_RTOR) [RM] 38.8.6 (pg 1251)
-  union {
-    struct {
-      uint32_t auto_baudrate : 1;
-      uint32_t send_break : 1;
-      uint32_t mute_mode : 1;
-      uint32_t flush_rx : 1;
-      uint32_t flush_tx : 1;
-      uint32_t reserved : 27;
-    } bitfield;
-    uint32_t full_reg;
-  } request;  // Request Register (USART_QRQ) [RM] 38.8.7 (pg 1252)
-  union {
-    struct {
-      const uint32_t parity_error : 1;
-      const uint32_t framing_error : 1;
-      const uint32_t noise_detection : 1;
-      const uint32_t overrun_error : 1;
-      const uint32_t idle : 1;
-      const uint32_t rx_not_empty : 1;
-      const uint32_t tx_complete : 1;
-      const uint32_t tx_empty : 1;
-      const uint32_t lin_break_detected : 1;
-      const uint32_t cts_interrupt : 1;
-      const uint32_t cts : 1;
-      const uint32_t rx_timeout : 1;
-      const uint32_t end_of_block : 1;
-      const uint32_t reserved : 1;
-      const uint32_t auto_baudrate_error : 1;
-      const uint32_t auto_baudrate_flag : 1;
-      const uint32_t busy : 1;
-      const uint32_t char_match : 1;
-      const uint32_t send_break : 1;
-      const uint32_t rx_wakeup : 1;
-      const uint32_t wakeup_from_stop : 1;
-      const uint32_t tx_enable_ack : 1;
-      const uint32_t rx_enable_ack : 1;
-      const uint32_t reserved2 : 2;
-      const uint32_t tx_complete_before_guard : 1;
-      const uint32_t reserved3 : 6;
-    } bitfield;
-    uint32_t full_reg;
-  } status;  // [RM] 38.8.8 Interrupt And Status Register (USART_ISR) (pg 1253)
-  union {
-    struct {
-      uint32_t parity_error_clear : 1;
-      uint32_t framing_error_clear : 1;
-      uint32_t noise_detected_clear : 1;
-      uint32_t overrun_clear : 1;
-      uint32_t clear_idle : 1;
-      uint32_t reserved : 1;
-      uint32_t tx_complete_clear : 1;
-      uint32_t tx_complete_before_guard_clear : 1;
-      uint32_t lin_break_detected_clear : 1;
-      uint32_t cts_clear : 1;
-      uint32_t reserved2 : 1;
-      uint32_t rx_timeout_clear : 1;
-      uint32_t end_of_block_clear : 1;
-      uint32_t reserved3 : 4;
-      uint32_t char_match_clear : 1;
-      uint32_t reserved4 : 2;
-      uint32_t wakeup_from_stop_clear : 1;
-      uint32_t reserved5 : 11;
-    } bitfield;
-    uint32_t full_reg;
-  } interrupt_clear;  // [RM] 38.8.9 Interrupt Flag Clear Register (USART_ICR)
-                      // (pg 1257)
-  uint32_t rx_data;
-  uint32_t tx_data;
-};
-typedef volatile UartStruct UartReg;
-inline UartReg *const Uart1Base = reinterpret_cast<UartReg *>(0x40013800);
-inline UartReg *const Uart2Base = reinterpret_cast<UartReg *>(0x40004400);
-inline UartReg *const Uart3Base = reinterpret_cast<UartReg *>(0x40004800);
-inline UartReg *const Uart4Base = reinterpret_cast<UartReg *>(0x40004C00);
-
 // [RM] 11.4.4 DMA channels (pg 302)
 enum class DmaChannel : uint8_t {
   Chan1 = 0,
@@ -394,51 +225,6 @@ inline void DmaClearInt(DmaReg *dma, DmaChannel chan, DmaInterrupt interrupt) {
   x <<= 4 * static_cast<uint8_t>(chan);  // 4 events per channel ([RM] 11.6.2)
   dma->interrupt_clear.full_reg = x;
 }
-
-// [RM] 40.6 SPI Registers (pg 1330)
-struct SpiStruct {
-  struct {
-    uint32_t clock_phase : 1;
-    uint32_t clock_polarity : 1;
-    uint32_t master : 1;
-    uint32_t bitrate : 3;
-    uint32_t enable : 1;
-    uint32_t lsb_first : 1;
-    uint32_t internal_slave_select : 1;
-    uint32_t sw_slave_management : 1;
-    uint32_t rx_only : 1;
-    uint32_t crc_length : 1;
-    uint32_t crc_next : 1;
-    uint32_t crc_enable : 1;
-    uint32_t bidir_output : 1;
-    uint32_t bidir_mode : 1;
-    uint32_t reserved : 16;
-  } control_reg1;  // Control register 1 (SPIx_CR1) [RM] 40.6.1 (pg 1330)
-  struct {
-    uint32_t rx_dma : 1;
-    uint32_t tx_dma : 1;
-    uint32_t ss_output : 1;
-    uint32_t nss_pulse : 1;
-    uint32_t frame_format : 1;
-    uint32_t error_interrupt : 1;
-    uint32_t rx_not_empty_interrupt : 1;
-    uint32_t tx_empty_interrupt : 1;
-    uint32_t data_size : 4;
-    uint32_t fifo_rx_threshold : 1;
-    uint32_t last_dma_rx : 1;
-    uint32_t last_dma_tx : 1;
-    uint32_t reserved : 17;
-  } control2;  // Control register 2 (SPIx_CR2) [RM] 40.6.2
-  uint32_t status;
-  uint32_t data;
-  uint32_t crc_polynomial;
-  uint32_t rx_crc;
-  uint32_t tx_crc;
-};
-typedef volatile SpiStruct SpiReg;
-inline SpiReg *const Spi1Base = reinterpret_cast<SpiReg *>(0x40013000);
-inline SpiReg *const Spi2Base = reinterpret_cast<SpiReg *>(0x40003800);
-inline SpiReg *const Spi3Base = reinterpret_cast<SpiReg *>(0x40003C00);
 
 // [RM] 37.7 I2C Registers
 struct I2CStruct {
